remove unused play button, create accessible_stack widget
