+++ lib/common/common.dart
@@ -88,3 +88,4 @@ export 'ui/custom/text_field.dart';
 export 'ui/custom/video_viewer.dart';
 export 'ui/custom/image_viewer.dart';
 export 'ui/custom/viewer.dart';
+export 'ui/custom/accessible_stack.dart';
